+++ libs/wingsdk/src/shared-aws/index.ts
@@ -1,0 +1,1 @@
+export * from "./types";
