+++ libs/wingsdk/src/shared-aws/permissions.ts
@@ -1,26 +1,5 @@
 import * as cloud from "../cloud";
-
-/**
- * AWS IAM Policy Statement.
- */
-export interface PolicyStatement {
-  /** Actions */
-  readonly actions?: string[];
-  /** Resources */
-  readonly resources?: string[];
-  /** Effect ("Allow" or "Deny") */
-  readonly effect?: Effect;
-}
-
-/**
- * The Effect element of an AWS IAM policy statement
- */
-export enum Effect {
-  /** Allow */
-  ALLOW = "Allow",
-  /** Deny */
-  DENY = "Deny",
-}
+import { PolicyStatement } from "./types";
 
 export function calculateTopicPermissions(
   arn: string,
